oops

+++ LZ2D_POST.py
@@ -174,18 +174,18 @@ def Score_train(i: int):
     # Data
     # tmax = 50
     traj_len = 64 
-    times = torch.arange(traj_len)
+    times = torch.arange(traj_len).float()
 
     simt = sim(N=config["points"], M=config["points"], noise=config["noise"])
     simt.init_observer(observer)
-    simt.data = load_data('train.h5').float()
+    simt.data = load_data('train.h5')
     simt.obs = simt.observe()
     simt.time = times[None,...].repeat(config["train_sim"],1)
     process_sim(simt)
 
     simv = sim(N=config["points"], M=config["points"], noise=config["noise"])
     simv.init_observer(observer)
-    simv.data = load_data('valid.h5').float()
+    simv.data = load_data('valid.h5')
     simv.obs = simv.observe()
     simv.time = times[None,...].repeat(config["val_sim"],1)
     process_sim(simv)
